move commandexecutor to pkg
