+++ pkg/helmutils/Helmutils.go
@@ -3,7 +3,7 @@ package helmutils
 import (
 	"context"
 
-	"github.com/asciich/asciichgolangpublic/commandexecutor"
+	"github.com/asciich/asciichgolangpublic/pkg/commandexecutor"
 	"github.com/asciich/asciichgolangpublic/pkg/helmutils/helminterfaces"
 	"github.com/asciich/asciichgolangpublic/pkg/helmutils/helmparameteroptions"
 )
