+++ pkg/kubernetesutils/kubernetesinterfaces/KubernetesCluster.go
@@ -3,7 +3,7 @@ package kubernetesinterfaces
 import (
 	"context"
 
-	"github.com/asciich/asciichgolangpublic/commandexecutor"
+	"github.com/asciich/asciichgolangpublic/pkg/commandexecutor"
 	"github.com/asciich/asciichgolangpublic/pkg/kubernetesutils/kubernetesimplementationindependend"
 	"github.com/asciich/asciichgolangpublic/pkg/kubernetesutils/kubernetesparameteroptions"
 )
